examples/runtime-and-ops/cold_start_mitigation/requirements.txt
azure-functions
azure-functions-logging-python
requests
